requirements.txt
autopep8==1.6.0
azure-common==1.1.28
azure-core==1.23.0
azure-functions==1.10.1
azure-identity==1.8.0
azure-servicebus==7.6.0
azure-storage-blob==12.11.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
et-xmlfile==1.1.0
google-api-core==2.7.1
google-auth==2.6.2
googleapis-common-protos==1.56.0
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.3
openpyxl==3.0.9
portalocker==2.4.0
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
typing_extensions==4.1.1
uamqp==1.5.3
urllib3==1.26.9
